fix infering max_seq_length from sparse inputs

+++ merlin/models/tf/blocks/retrieval/base.py
@@ -292,7 +292,7 @@ class ItemRetrievalScorer(Block):
 
     def rescore_false_negatives(self, positive_item_ids, neg_samples_item_ids, negative_scores):
         # Removing dimensions of size 1 from the shape of the item ids, if applicable
-        positive_item_ids = tf.cast(tf.squeeze(positive_item_ids), tf.int32)
+        positive_item_ids = tf.cast(tf.squeeze(positive_item_ids), neg_samples_item_ids.dtype)
         neg_samples_item_ids = tf.squeeze(neg_samples_item_ids)
 
         # Reshapes positive and negative ids so that false_negatives_mask matches the scores shape
